address keyValues map and add documentation
